+++ daemons/dss-notify-v2/app.py
@@ -16,8 +16,9 @@ sys.path.insert(0, pkg_root)  # noqa
 import dss
 from dss import Config, Replica
 from dss.logging import configure_lambda_logging
-from dss.events.handlers.notify_v2 import (should_notify, notify_or_queue, notify, build_bundle_metadata_document,
-                                           build_deleted_bundle_metadata_document)
+from dss.events import get_bundle_metadata_document, get_deleted_bundle_metadata_document, record_event_for_bundle
+from dss.events.handlers.notify_v2 import should_notify, notify_or_queue, notify
+
 from dss.events.handlers.sync import exists
 from dss.subscriptions_v2 import get_subscription, get_subscriptions_for_replica
 
@@ -91,12 +92,12 @@ def launch_from_notification_queue(event, context):
         subscription = get_subscription(replica, owner, uuid)
         if subscription is not None:
             if "DELETE" == event_type:
-                metadata_document = build_deleted_bundle_metadata_document(key)
+                metadata_document = get_deleted_bundle_metadata_document(key)
             else:
                 if not exists(replica, key):
                     logger.warning(f"Key %s not found in replica %s, unable to notify %s", key, replica.name, uuid)
                     return
-                metadata_document = build_bundle_metadata_document(replica, key)
+                metadata_document = get_bundle_metadata_document(replica, key)
             if not notify(subscription, metadata_document, key):
                 # Erroring causes the message to remain in the queue
                 raise DSSFailedNotificationDelivery()
@@ -105,10 +106,10 @@ def launch_from_notification_queue(event, context):
 
 def _notify_subscribers(replica: Replica, key: str, is_delete_event: bool):
     if is_delete_event:
-        metadata_document = build_deleted_bundle_metadata_document(key)
+        metadata_document = get_deleted_bundle_metadata_document(key)
     else:
         if exists(replica, key):
-            metadata_document = build_bundle_metadata_document(replica, key)
+            metadata_document = record_event_for_bundle(replica, key)
         else:
             logger.error(f"Key %s not found in replica %s, unable to notify subscribers", key, replica.name)
             return
