+++ requirements-dev.txt
@@ -2,16 +2,16 @@
 asn1crypto==0.24.0
 attrs==19.1.0
 aws-xray-sdk==2.4.2
-awscli==1.16.229
+awscli==1.16.242
 azure-common==1.1.23
 azure-nspkg==3.0.2
 azure-storage==0.36.0
-boto3==1.9.219
-botocore==1.12.219
+boto3==1.9.232
+botocore==1.12.232
 cachetools==3.1.1
-certifi==2019.6.16
+certifi==2019.9.11
 cffi==1.12.3
-chalice==1.11.0
+chalice==1.11.1
 chardet==3.0.4
 click==6.7
 clickclick==1.2.2
@@ -22,16 +22,17 @@ coverage==4.5.4
 crc32c==1.7
 crcmod==1.7
 cryptography==2.7
-dcplib==3.3.0
+dcplib==3.7.0
 docutils==0.15.2
 domovoi==2.0.2
 elasticsearch==5.5.3
 elasticsearch-dsl==5.4.0
 entrypoints==0.3
 enum-compat==0.0.2
-Faker==2.0.1
+Faker==2.0.2
 fasteners==0.15
 flake8==3.7.8
+flash-flood==0.2.0
 Flask==1.1.1
 furl==2.0.0
 future==0.17.1
@@ -41,8 +42,8 @@ google-apitools==0.5.30
 google-auth==1.6.3
 google-auth-httplib2==0.0.3
 google-cloud-core==1.0.3
-google-cloud-storage==1.19.0
-google-resumable-media==0.3.3
+google-cloud-storage==1.19.1
+google-resumable-media==0.4.1
 googleapis-common-protos==1.6.0
 httpie==1.0.3
 httplib2==0.13.1
@@ -65,7 +66,7 @@ oauth2client==4.1.3
 orderedmultidict==1.0.1
 protobuf==3.9.1
 puremagic==1.4
-pyasn1==0.4.6
+pyasn1==0.4.7
 pyasn1-modules==0.2.6
 pycodestyle==2.5.0
 pycparser==2.19
@@ -74,7 +75,7 @@ Pygments==2.4.2
 PyJWT==1.7.1
 python-dateutil==2.8.0
 pytz==2019.2
-PyYAML==5.1
+PyYAML==5.1.2
 requests==2.22.0
 requests-aws4auth==0.9
 requests-http-signature==0.1.0
@@ -84,13 +85,13 @@ s3transfer==0.2.1
 six==1.12.0
 swagger-spec-validator==2.4.3
 termcolor==1.1.0
-text-unidecode==1.2
+text-unidecode==1.3
 typed-ast==1.4.0
 typing==3.6.4
 typing-extensions==3.7.4
 uritemplate==3.0.0
-urllib3==1.25.3
-Werkzeug==0.15.5
+urllib3==1.25.4
+Werkzeug==0.16.0
 wrapt==1.11.2
 xmltodict==0.12.0
 yq==2.7.2
