+++ requirements.txt
@@ -4,10 +4,10 @@ aws-xray-sdk==2.4.2
 azure-common==1.1.23
 azure-nspkg==3.0.2
 azure-storage==0.36.0
-boto3==1.9.219
-botocore==1.12.219
+boto3==1.9.232
+botocore==1.12.232
 cachetools==3.1.1
-certifi==2019.6.16
+certifi==2019.9.11
 cffi==1.12.3
 chardet==3.0.4
 Click==7.0
@@ -16,11 +16,12 @@ cloud-blobstore==3.2.0
 connexion==1.5.3
 crc32c==1.7
 cryptography==2.7
-dcplib==3.3.0
+dcplib==3.7.0
 docutils==0.15.2
 elasticsearch==5.5.3
 elasticsearch-dsl==5.4.0
 fasteners==0.15
+flash-flood==0.2.0
 Flask==1.1.1
 furl==2.0.0
 future==0.17.1
@@ -30,8 +31,8 @@ google-apitools==0.5.30
 google-auth==1.6.3
 google-auth-httplib2==0.0.3
 google-cloud-core==1.0.3
-google-cloud-storage==1.19.0
-google-resumable-media==0.3.3
+google-cloud-storage==1.19.1
+google-resumable-media==0.4.1
 googleapis-common-protos==1.6.0
 httplib2==0.13.1
 idna==2.8
@@ -50,13 +51,13 @@ oauth2client==4.1.3
 orderedmultidict==1.0.1
 protobuf==3.9.1
 puremagic==1.4
-pyasn1==0.4.6
+pyasn1==0.4.7
 pyasn1-modules==0.2.6
 pycparser==2.19
 PyJWT==1.7.1
 python-dateutil==2.8.0
 pytz==2019.2
-PyYAML==5.1
+PyYAML==5.1.2
 requests==2.22.0
 requests-aws4auth==0.9
 requests-http-signature==0.1.0
@@ -66,6 +67,6 @@ six==1.12.0
 swagger-spec-validator==2.4.3
 termcolor==1.1.0
 uritemplate==3.0.0
-urllib3==1.25.3
-Werkzeug==0.15.5
+urllib3==1.25.4
+Werkzeug==0.16.0
 wrapt==1.11.2
