+++ requirements.txt.in
@@ -20,4 +20,5 @@ urllib3 >= 1.21.1
 requests-http-signature >= 0.0.3
 aws-xray-sdk >= 1.0
 pyjwt >= 1.6.4
-pyyaml >= 4.2b1, <= 5.1.0
+pyyaml >= 4.2b1
+flash-flood >= 0.2.0
